requirements.txt
python-dotenv>=1.0.0
streamlit>=1.30.0
streamlit-ace>=0.1.1
openai>=1.0.0
langchain>=0.1.0 
langchain-openai>=0.3.0